Complete rest basic API

+++ src/main/java/com/vietphuongdo/shopapp/entities/OrderDetail.java
@@ -25,13 +25,13 @@ public class OrderDetail {
     private Product product;
 
     @Column(name = "price",nullable = false)
-    private int price;
+    private Float price;
 
     @Column(name = "number_of_products",nullable = false)
     private int numberOfProducts;
 
     @Column(name = "total_money",nullable = false)
-    private int totalMoney;
+    private Float totalMoney;
 
     @Column(name = "color")
     private String color;
